requirements.txt
flask==3.0.3
flask-cors==4.0.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python-headless==4.10.0.84
ultralytics==8.2.62
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1